requirements/dev.txt
# Additional requirements useful for development.

-r base.txt

mock==1.0b1

django-webtest==1.5.3
WebTest==1.3.4
WebOb==1.2

coverage==3.5.2

pytest==2.2.4
py==1.4.9
pytest-django==1.4
pytest-cov==1.6
cov-core==1.6
